chore: basic seed data txt files
