+++ .gitignore
@@ -30,8 +30,6 @@
 yarn-debug.log*
 .yarn-integrity
 
-/seeders/data
-
 # Ignore Coveralls coverage reports
 /coverage
 .env
