+++ app/seeders/data/soft-skills.txt
@@ -1,0 +1,9 @@
+Can you tell me about a time when you worked as part of a group?
+Can you tell me about a time when you had to ask for help?
+Tell me about a time when you were asked to do something you had never done before. How did you react? What did you learn?
+What are the three things that are most important to you in a job?
+Give an example of when you had to work with someone who was difficult to get along with. How did you handle interactions with that person?
+Tell me about the last time something significant didn’t go according to plan at work. What was your role? What was the outcome?
+In your last job, how do you think your co-workers would describe working with you?
+Tell me about your proudest professional accomplishment.
+Tell me about a time you were dissatisfied in your work. What could have been done to make it better?
