+++ app/seeders/default_list_seeder.rb
@@ -1,0 +1,20 @@
+# class DefaultListSeeder
+
+#   # LISTS = [
+#   #   {name: "Soft Skills",user: User.first}
+#   # ]
+
+#   def self.seed!
+#     list = {name: "Soft Skills",user: User.first}
+#     list.save!
+
+#     list_questions = {list: List.first, question: }
+#     # LISTS.each do |item|
+#     #   list = List.find_or_initialize_by(name: item[:name], user: item[:user])
+#     #   list.save!
+#     # end
+
+#   end
+
+
+# end
